many minor primitive cast changes to remove compiler lossy warnings

+++ src/algorithms/compGeometry/ParabolaLeastSquares.java
@@ -187,9 +187,9 @@ public class ParabolaLeastSquares {
         for (int i = (3 - 1); i >= 0; i--) {
             c[i] = (float)qTY[i];
             for (int j = (i + 1); j < 3; j++) {
-                c[i] -= c[j] * r.get(i, j);
+                c[i] -= c[j] * (float)r.get(i, j);
             }
-            c[i] /= r.get(i, i);
+            c[i] /= (float)r.get(i, i);
         }
         
         if (debug) {
